widu-server for fiji/server-folder/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.40.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
keras==2.11.0
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
numpy==1.24.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
pillow==10.4.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-gpu==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
